Trait cs431::lock::RawLock

pub unsafe trait RawLock: Default + Send + Sync {
    type Token;

    // Required methods
    fn lock(&self) -> Self::Token;
    unsafe fn unlock(&self, token: Self::Token);
}

Raw lock interface.

Safety

Implementations of this trait must ensure that the lock is actually exclusive: a lock can’t be acquired while the lock is already locked.

Required Associated Types

type Token

Raw lock’s token type.

Required Methods

fn lock(&self) -> Self::Token

Acquires the raw lock.

unsafe fn unlock(&self, token: Self::Token)

Releases the raw lock.

Safety

unlock() should be called with the token given by the corresponding lock().

Object Safety

This trait is not object safe.