Trait cs431::lock::RawTryLock

pub unsafe trait RawTryLock: RawLock {
    // Required method
    fn try_lock(&self) -> Result<Self::Token, ()>;
}

Raw lock interface for the try_lock API.

Safety

Implementations of this trait must ensure that the lock is actually exclusive: a lock can’t be acquired while the lock is already locked. Also, try_lock(), when successful, should return a token that can be used for RawLock::unlock.

Required Methods

fn try_lock(&self) -> Result<Self::Token, ()>

Tries to acquire the raw lock.

Object Safety

This trait is not object safe.