cs431::lockfree::list

Struct Cursor

pub struct Cursor<'g, K, V> { /* private fields */ }

Linked list cursor.

Implementations

impl<'g, K, V> Cursor<'g, K, V>
where K: Ord,

pub fn new(prev: &'g Atomic<Node<K, V>>, curr: Shared<'g, Node<K, V>>) -> Self

Creates a cursor.

pub fn curr(&self) -> Shared<'g, Node<K, V>>

Returns the current node.

pub fn find_harris(&mut self, key: &K, guard: &'g Guard) -> Result<bool, ()>

Clean up a chain of logically removed nodes in each traversal.

pub fn find_harris_michael( &mut self, key: &K, guard: &'g Guard, ) -> Result<bool, ()>

Clean up a single logically removed node in each traversal.

pub fn find_harris_herlihy_shavit( &mut self, key: &K, guard: &'g Guard, ) -> Result<bool, ()>

Doesnโt preform any cleanup. Gotta go fast. Doesnโt fail.

pub fn lookup(&self) -> &'g V

Lookups the value at the current node.

Panics

Panics if the current node is null.

pub fn insert( &mut self, node: Owned<Node<K, V>>, guard: &'g Guard, ) -> Result<(), Owned<Node<K, V>>>

Inserts a value between the previous and current node.

pub fn delete(&mut self, guard: &'g Guard) -> Result<&'g V, ()>

Deletes the current node.

Panics

Panics if the current node is null.