cs431::lock

Trait RawLock

pub unsafe trait RawLock:
    Default
    + Send
    + Sync {
    type Token;

    // Required methods
    fn lock(&self) -> Self::Token;
    unsafe fn unlock(&self, token: Self::Token);
}

Raw lock interface.

Safety

Implementations of this trait must ensure that the lock is actually exclusive: a lock can’t be acquired while the lock is already locked.

Required Associated Types

type Token

Raw lock’s token type.

We don’t enforce Send/Sync, as some locks may not satisfy it. We restrict them at Send/Sync impl for LockGuard.

Required Methods

fn lock(&self) -> Self::Token

Acquires the raw lock.

unsafe fn unlock(&self, token: Self::Token)

Releases the raw lock.

Safety

self must be an acquired lock.

token must be from a RawLock::lock or RawTryLock::try_lock call to self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.